Add testsolve controller, organize access_controller, organize twig files
