+++ config/packages/security.yaml
@@ -47,20 +47,7 @@ security:
   # Note: Only the *first* access control that matches will be used
   access_control:
     - { path: ^/(%app.allowed_locales%)/auth, roles: PUBLIC_ACCESS }
-    - {
-        path: ^/(%app.allowed_locales%)/test/introduction/\d+$,
-        roles: PUBLIC_ACCESS,
-      }
-    - {
-        path: ^/(%app.allowed_locales%)/test/conclusion/\d+$,
-        roles: PUBLIC_ACCESS,
-      }
-    - {
-        path: ^/(%app.allowed_locales%)/test/video/\d+/\d+$,
-        roles: PUBLIC_ACCESS,
-      }
-    - { path: ^/(%app.allowed_locales%)/test/notFound, roles: PUBLIC_ACCESS }
-    - { path: ^/(%app.allowed_locales%)/test/notValid, roles: PUBLIC_ACCESS }
+    - { path: ^/(%app.allowed_locales%)/testSolve, roles: PUBLIC_ACCESS }
     - { path: ^/(%app.allowed_locales%), roles: IS_AUTHENTICATED_FULLY }
 
 when@test:
