+++ src/DataTable/Type/TestDataTableType.php
@@ -46,7 +46,7 @@ class TestDataTableType extends AbstractDataTableType
                                 [
                                     'label' => 'data_table.test.test',
                                     'href' => function(Test $test): string {
-                                        return $this->urlGenerator->generate('app_test_introduction', [
+                                        return $this->urlGenerator->generate('app_testsolve_introduction', [
                                             '_locale' => $test->getModule()->getLanguage(),
                                             'id' => $test->getId()
                                         ]);
