+++ templates/test/index.html.twig
@@ -1,9 +1,5 @@
 {% extends 'base.authenticated.html.twig' %}
 
-{% block title %} 
-    {{ 'templates.test.index.title'|trans }} 
-{% endblock %}
-
 {% block container %}
     {{ data_table(test_data_table) }}
 {% endblock %}
