--- templates/test/solve.notFound.html.twig
@@ -1,12 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %} 
-    {{ 'templates.test.solve.notFound.title'|trans }} 
-{% endblock %}
-
-{% block body %}
-    {{ component('TestErrorSolve', {
-        header: 'templates.test.notFound.header',
-        content: 'templates.test.notFound.content'
-    }) }}
-{% endblock %}
