--- templates/test/solve.notValid.html.twig
@@ -1,12 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %} 
-    {{ 'templates.test.solve.notValid.title'|trans }} 
-{% endblock %}
-
-{% block body %}
-    {{ component('TestErrorSolve', {
-        header: 'templates.test.notValid.header',
-        content: 'templates.test.notValid.content'
-    }) }}
-{% endblock %}
