+++ templates/testSolve/conclusion.html.twig
@@ -1,4 +1,4 @@
-{% extends 'test/base.solve.html.twig' %}
+{% extends 'testSolve/base.html.twig' %}
 
 {% block wrapper %}
     <div class="d-flex flex-column justify-content-center">
